+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -463,7 +463,7 @@ const VirtualListBaseFactory = (type) => {
 
 				this.lastFocusedIndex = nextIndex;
 
-				// VirtualGridList does not support different item size yet. So we don't need to consider the condition related with it.
+				// VirtualGridList does not support different item size yet.
 				if (isNextItemInView || !this.props.itemSizes && (numOfItemsInPage !== dimensionToExtent || row === nextRow)) {
 					// The next item could be still out of viewport. So we need to prevent scrolling into view with `isScrolledBy5way` flag.
 					this.isScrolledBy5way = true;